AdicionandoPonto

+++ public/js/calc_functions.js
@@ -14,13 +14,13 @@ var valor1 = 0
 function atualizarOperacao(btn){
     const display = document.getElementById('display');
     operador = btn.value;
-    valor1 = parseInt(display.value);
+    valor1 = parseFloat(display.value);
     display.value = '0';
 }
 
 function calcularOperacao(){
     const display = document.getElementById('display');
-    const valor2 = parseInt(display.value);
+    const valor2 = parseFloat(display.value);
     valor1 = eval(valor1+operador+valor2);
     display.value = valor1;
     operador = '';
@@ -30,3 +30,5 @@ function manipularTeclado(){
     if(/[0-9]/.test(event.key))
         atualizarDisplay({value: event.key});
 }
+
+
